pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragfin"
version = "0.16.0"
description = "Local-first financial report assistant (RAG-Fin)"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
  "pypdf>=4.2.0",
  "chromadb>=0.5.0",
  "sentence-transformers>=2.7.0",
  "streamlit>=1.35.0",
  "pandas>=2.2.0",
  "rank-bm25>=0.2.2",
  "PyYAML>=6.0.1",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[project.scripts]
ragfin = "ragfin.cli:main"
ragfin-ui = "ragfin.ui:main"
ragfin-api = "ragfin.api:main"
ragfin-demo = "ragfin.demo:main"
ragfin-doctor = "ragfin.doctor:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "ragfin*", "app*"]
